pkg: move proposal to primitives
