+++ lib/bmultisig-browser.js
@@ -24,5 +24,6 @@ bmultisig.Client = require('./client');
 
 // primitives
 bmultisig.Cosigner = require('./primitives/cosigner');
+bmultisig.Proposal = require('./primitives/proposal');
 
 bmultisig.pkg = require('./pkg');
