+++ lib/bmultisig.js
@@ -44,7 +44,9 @@ bmultisig.define('Plugin', './plugin');
 // multisig http client
 bmultisig.define('Client', './client');
 
+// primitives
 bmultisig.define('Cosigner', './primitives/cosigner');
+bmultisig.define('Proposal', './primitives/proposal');
 
 bmultisig.define('MultisigDB', './multisigdb');
 bmultisig.define('ProposalDB', './proposaldb');
