+++ lib/primitives/proposal.js
@@ -10,8 +10,8 @@ const bcoin = require('bcoin');
 const {Struct} = require('bufio');
 const {common} = bcoin.wallet;
 const {Outpoint, TX} = bcoin;
-const Cosigner = require('./primitives/cosigner');
-const layout = require('./layout').proposaldb;
+const Cosigner = require('./cosigner');
+const layout = require('../layout').proposaldb;
 
 /**
  * Proposal status
