+++ lib/proposaldb.js
@@ -8,7 +8,7 @@
 
 const assert = require('assert');
 const {Outpoint, CoinView, MTX} = require('bcoin');
-const Proposal = require('./proposal');
+const Proposal = require('./primitives/proposal');
 const {MapLock, Lock} = require('bmutex');
 const layout = require('./layout').proposaldb;
