+++ test/proposals-test.js
@@ -13,7 +13,7 @@ const WalletDB = bcoin.wallet.WalletDB;
 const WalletNodeClient = require('../lib/walletclient');
 const MultisigDB = require('../lib/multisigdb');
 const Cosigner = require('../lib/primitives/cosigner');
-const Proposal = require('../lib/proposal');
+const Proposal = require('../lib/primitives/proposal');
 
 const TEST_XPUB_PATH = 'm/44\'/0\'/0\'';
 const TEST_WALLET_ID = 'test';
